Split oversized test files into submodules
tree/tests.rs (1152 lines) split into:
  - tests/node.rs (95 lines) - TreeNode tests
  - tests/state.rs (175 lines) - TreeState tests
  - tests/component.rs (570 lines) - update, view, integration tests
  - tests/events.rs (311 lines) - handle_event, dispatch_event, disabled

multi_progress/tests.rs (1088 lines) split into:
  - tests/item.rs (64 lines) - ProgressItemStatus, ProgressItem tests
  - tests/state.rs (240 lines) - state creation, manipulation, counting
  - tests/component.rs (550 lines) - update, scroll, view, coverage tests
  - tests/events.rs (239 lines) - handle_event, dispatch_event, disabled

All 2049 unit tests pass. Snapshot files renamed for new module paths.
