+++ src/component/multi_progress/tests/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+use crate::input::{Event, KeyCode};
+
+mod component;
+mod events;
+mod item;
+mod state;
