+++ src/component/tree/tests/component.rs
@@ -1,279 +1,4 @@
 use super::*;
-use crate::input::{Event, KeyCode};
-
-// TreeNode tests
-
-#[test]
-fn test_node_new() {
-    let node = TreeNode::new("Label", "data");
-    assert_eq!(node.label(), "Label");
-    assert_eq!(node.data(), &"data");
-    assert!(!node.is_expanded());
-    assert!(!node.has_children());
-}
-
-#[test]
-fn test_node_new_expanded() {
-    let node: TreeNode<()> = TreeNode::new_expanded("Label", ());
-    assert!(node.is_expanded());
-}
-
-#[test]
-fn test_node_set_label() {
-    let mut node = TreeNode::new("Old", ());
-    node.set_label("New");
-    assert_eq!(node.label(), "New");
-}
-
-#[test]
-fn test_node_data_mut() {
-    let mut node = TreeNode::new("Label", 42);
-    *node.data_mut() = 100;
-    assert_eq!(node.data(), &100);
-}
-
-#[test]
-fn test_node_add_child() {
-    let mut parent = TreeNode::new("Parent", ());
-    parent.add_child(TreeNode::new("Child 1", ()));
-    parent.add_child(TreeNode::new("Child 2", ()));
-
-    assert!(parent.has_children());
-    assert_eq!(parent.children().len(), 2);
-}
-
-#[test]
-fn test_node_children_mut() {
-    let mut parent = TreeNode::new("Parent", ());
-    parent.add_child(TreeNode::new("Child", ()));
-
-    parent.children_mut()[0].set_label("Modified");
-    assert_eq!(parent.children()[0].label(), "Modified");
-}
-
-#[test]
-fn test_node_expand_collapse() {
-    let mut node = TreeNode::new("Node", ());
-    assert!(!node.is_expanded());
-
-    node.expand();
-    assert!(node.is_expanded());
-
-    node.collapse();
-    assert!(!node.is_expanded());
-}
-
-#[test]
-fn test_node_toggle() {
-    let mut node = TreeNode::new("Node", ());
-    assert!(!node.is_expanded());
-
-    node.toggle();
-    assert!(node.is_expanded());
-
-    node.toggle();
-    assert!(!node.is_expanded());
-}
-
-#[test]
-fn test_node_set_expanded() {
-    let mut node = TreeNode::new("Node", ());
-    node.set_expanded(true);
-    assert!(node.is_expanded());
-    node.set_expanded(false);
-    assert!(!node.is_expanded());
-}
-
-#[test]
-fn test_node_clone() {
-    let mut node = TreeNode::new("Parent", "data");
-    node.add_child(TreeNode::new("Child", "child_data"));
-    node.expand();
-
-    let cloned = node.clone();
-    assert_eq!(cloned.label(), "Parent");
-    assert_eq!(cloned.data(), &"data");
-    assert!(cloned.is_expanded());
-    assert_eq!(cloned.children().len(), 1);
-}
-
-// TreeState tests
-
-#[test]
-fn test_state_new() {
-    let roots = vec![TreeNode::new("Root", ())];
-    let state = TreeState::new(roots);
-
-    assert_eq!(state.roots().len(), 1);
-    assert_eq!(state.selected_index(), Some(0));
-    assert!(!state.is_empty());
-}
-
-#[test]
-fn test_state_new_empty() {
-    let state: TreeState<()> = TreeState::new(Vec::new());
-    assert!(state.is_empty());
-    assert_eq!(state.selected_index(), None);
-}
-
-#[test]
-fn test_state_default() {
-    let state: TreeState<()> = TreeState::default();
-    assert!(state.is_empty());
-    assert_eq!(state.selected_index(), None);
-}
-
-#[test]
-fn test_state_set_roots() {
-    let mut state: TreeState<i32> = TreeState::default();
-    state.set_roots(vec![TreeNode::new("Root", 1)]);
-
-    assert_eq!(state.roots().len(), 1);
-    assert_eq!(state.selected_index(), Some(0));
-}
-
-#[test]
-fn test_state_set_roots_to_empty() {
-    let mut state = TreeState::new(vec![TreeNode::new("Root", ())]);
-    state.set_roots(Vec::new());
-    assert_eq!(state.selected_index(), None);
-}
-
-#[test]
-fn test_state_roots_mut() {
-    let mut state = TreeState::new(vec![TreeNode::new("Root", ())]);
-    state.roots_mut()[0].set_label("Modified");
-
-    assert_eq!(state.roots()[0].label(), "Modified");
-}
-
-#[test]
-fn test_state_flatten_single() {
-    let state = TreeState::new(vec![TreeNode::new("Root", ())]);
-    let flat = state.flatten();
-
-    assert_eq!(flat.len(), 1);
-    assert_eq!(flat[0].label, "Root");
-    assert_eq!(flat[0].depth, 0);
-}
-
-#[test]
-fn test_state_flatten_with_children() {
-    let mut root = TreeNode::new_expanded("Root", ());
-    root.add_child(TreeNode::new("Child 1", ()));
-    root.add_child(TreeNode::new("Child 2", ()));
-
-    let state = TreeState::new(vec![root]);
-    let flat = state.flatten();
-
-    assert_eq!(flat.len(), 3);
-    assert_eq!(flat[0].label, "Root");
-    assert_eq!(flat[0].depth, 0);
-    assert_eq!(flat[1].label, "Child 1");
-    assert_eq!(flat[1].depth, 1);
-    assert_eq!(flat[2].label, "Child 2");
-    assert_eq!(flat[2].depth, 1);
-}
-
-#[test]
-fn test_state_flatten_collapsed() {
-    let mut root = TreeNode::new("Root", ()); // Not expanded
-    root.add_child(TreeNode::new("Child", ()));
-
-    let state = TreeState::new(vec![root]);
-    let flat = state.flatten();
-
-    // Children not visible when collapsed
-    assert_eq!(flat.len(), 1);
-}
-
-#[test]
-fn test_state_selected_path() {
-    let mut root = TreeNode::new_expanded("Root", ());
-    root.add_child(TreeNode::new("Child", ()));
-
-    let mut state = TreeState::new(vec![root]);
-    assert_eq!(state.selected_path(), Some(vec![0]));
-
-    state.selected_index = Some(1);
-    assert_eq!(state.selected_path(), Some(vec![0, 0]));
-}
-
-#[test]
-fn test_state_selected_path_empty() {
-    let state: TreeState<()> = TreeState::new(Vec::new());
-    assert_eq!(state.selected_path(), None);
-}
-
-#[test]
-fn test_state_selected_node() {
-    let mut root = TreeNode::new_expanded("Root", "root_data");
-    root.add_child(TreeNode::new("Child", "child_data"));
-
-    let mut state = TreeState::new(vec![root]);
-
-    let selected = state.selected_node();
-    assert!(selected.is_some());
-    assert_eq!(selected.unwrap().data(), &"root_data");
-
-    state.selected_index = Some(1);
-    let selected = state.selected_node();
-    assert_eq!(selected.unwrap().data(), &"child_data");
-}
-
-#[test]
-fn test_state_selected_node_empty() {
-    let state: TreeState<()> = TreeState::new(Vec::new());
-    assert!(state.selected_node().is_none());
-}
-
-#[test]
-fn test_state_expand_all() {
-    let mut root = TreeNode::new("Root", ());
-    let mut child = TreeNode::new("Child", ());
-    child.add_child(TreeNode::new("Grandchild", ()));
-    root.add_child(child);
-
-    let mut state = TreeState::new(vec![root]);
-    assert_eq!(state.visible_count(), 1);
-
-    state.expand_all();
-    assert_eq!(state.visible_count(), 3);
-}
-
-#[test]
-fn test_state_collapse_all() {
-    let mut root = TreeNode::new_expanded("Root", ());
-    let mut child = TreeNode::new_expanded("Child", ());
-    child.add_child(TreeNode::new("Grandchild", ()));
-    root.add_child(child);
-
-    let mut state = TreeState::new(vec![root]);
-    assert_eq!(state.visible_count(), 3);
-
-    state.collapse_all();
-    assert_eq!(state.visible_count(), 1);
-    assert_eq!(state.selected_index(), Some(0));
-}
-
-#[test]
-fn test_state_collapse_all_empty() {
-    let mut state: TreeState<()> = TreeState::new(Vec::new());
-    state.collapse_all();
-    assert_eq!(state.selected_index(), None);
-}
-
-#[test]
-fn test_state_visible_count() {
-    let mut root = TreeNode::new_expanded("Root", ());
-    root.add_child(TreeNode::new("Child 1", ()));
-    root.add_child(TreeNode::new("Child 2", ()));
-
-    let state = TreeState::new(vec![root]);
-    assert_eq!(state.visible_count(), 3);
-}
-
-// Tree component tests
 
 #[test]
 fn test_init() {
@@ -842,311 +567,4 @@ fn test_unicode_node_labels() {
     // Should navigate through unicode-labeled nodes without issue
     assert_eq!(state.selected_index(), Some(1));
     assert_eq!(state.selected_node().unwrap().label(), "설정");
-}
-
-// ========== handle_event Tests ==========
-
-fn make_tree_state() -> TreeState<&'static str> {
-    let mut root = TreeNode::new_expanded("Root", "root");
-    root.add_child(TreeNode::new("Child 1", "child1"));
-    root.add_child(TreeNode::new("Child 2", "child2"));
-    TreeState::new(vec![root])
-}
-
-#[test]
-fn test_handle_event_up_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-    state.selected_index = Some(1);
-
-    let event = Event::key(KeyCode::Up);
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Up));
-}
-
-#[test]
-fn test_handle_event_down_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let event = Event::key(KeyCode::Down);
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Down));
-}
-
-#[test]
-fn test_handle_event_expand_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let event = Event::key(KeyCode::Right);
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Expand));
-}
-
-#[test]
-fn test_handle_event_collapse_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let event = Event::key(KeyCode::Left);
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Collapse));
-}
-
-#[test]
-fn test_handle_event_toggle_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let event = Event::char(' ');
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Toggle));
-}
-
-#[test]
-fn test_handle_event_select_when_focused() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let event = Event::key(KeyCode::Enter);
-    let msg = Tree::<&str>::handle_event(&state, &event);
-    assert_eq!(msg, Some(TreeMessage::Select));
-}
-
-#[test]
-fn test_handle_event_vim_keys() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    let msg_k = Tree::<&str>::handle_event(&state, &Event::char('k'));
-    assert_eq!(msg_k, Some(TreeMessage::Up));
-
-    let msg_j = Tree::<&str>::handle_event(&state, &Event::char('j'));
-    assert_eq!(msg_j, Some(TreeMessage::Down));
-
-    let msg_h = Tree::<&str>::handle_event(&state, &Event::char('h'));
-    assert_eq!(msg_h, Some(TreeMessage::Collapse));
-
-    let msg_l = Tree::<&str>::handle_event(&state, &Event::char('l'));
-    assert_eq!(msg_l, Some(TreeMessage::Expand));
-}
-
-#[test]
-fn test_handle_event_ignored_when_unfocused() {
-    let state = make_tree_state();
-    // focused is false by default
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Down));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Enter));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::char('j'));
-    assert_eq!(msg, None);
-}
-
-// ========== dispatch_event Tests ==========
-
-#[test]
-fn test_dispatch_event() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-
-    // Dispatch Down: should move selection from 0 to 1
-    let output = Tree::<&str>::dispatch_event(&mut state, &Event::key(KeyCode::Down));
-    assert_eq!(output, None); // Down returns None but updates state
-    assert_eq!(state.selected_index(), Some(1));
-
-    // Dispatch Enter: should select the current node
-    let output = Tree::<&str>::dispatch_event(&mut state, &Event::key(KeyCode::Enter));
-    assert_eq!(output, Some(TreeOutput::Selected(vec![0, 0])));
-}
-
-// ========== Instance Method Tests ==========
-
-#[test]
-fn test_instance_methods() {
-    let mut state = make_tree_state();
-
-    // is_focused / set_focused
-    assert!(!state.is_focused());
-    state.set_focused(true);
-    assert!(state.is_focused());
-    state.set_focused(false);
-    assert!(!state.is_focused());
-
-    // dispatch_event via instance method
-    state.set_focused(true);
-    let output = state.dispatch_event(&Event::key(KeyCode::Down));
-    assert_eq!(output, None); // Down returns None but updates state
-    assert_eq!(state.selected_index(), Some(1));
-
-    // update via instance method
-    let output = state.update(TreeMessage::Select);
-    assert_eq!(output, Some(TreeOutput::Selected(vec![0, 0])));
-
-    // handle_event via instance method
-    let msg = state.handle_event(&Event::key(KeyCode::Up));
-    assert_eq!(msg, Some(TreeMessage::Up));
-}
-
-#[test]
-fn test_selected_item() {
-    let state = TreeState::new(vec![TreeNode::new("root", "data")]);
-    assert_eq!(state.selected_item().unwrap().data(), &"data");
-    assert_eq!(state.selected_item().unwrap().label(), state.selected_node().unwrap().label());
-}
-
-// ========== Disabled State Tests ==========
-
-#[test]
-fn test_disabled_default_false() {
-    let state = TreeState::new(vec![TreeNode::new("Root", ())]);
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_set_disabled() {
-    let mut state = TreeState::new(vec![TreeNode::new("Root", ())]);
-    assert!(!state.is_disabled());
-
-    state.set_disabled(true);
-    assert!(state.is_disabled());
-
-    state.set_disabled(false);
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_with_disabled() {
-    let state = TreeState::new(vec![TreeNode::new("Root", ())]).with_disabled(true);
-    assert!(state.is_disabled());
-
-    let state = TreeState::new(vec![TreeNode::new("Root", ())]).with_disabled(false);
-    assert!(!state.is_disabled());
-}
-
-#[test]
-fn test_handle_event_ignored_when_disabled() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-    state.set_disabled(true);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Down));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Up));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Enter));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Left));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::key(KeyCode::Right));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::char(' '));
-    assert_eq!(msg, None);
-
-    let msg = Tree::<&str>::handle_event(&state, &Event::char('j'));
-    assert_eq!(msg, None);
-}
-
-#[test]
-fn test_update_ignored_when_disabled() {
-    let mut root = TreeNode::new("Root", ());
-    root.add_child(TreeNode::new("Child", ()));
-
-    let mut state = TreeState::new(vec![root]);
-    state.set_disabled(true);
-
-    // Down should be ignored
-    let output = Tree::update(&mut state, TreeMessage::Down);
-    assert_eq!(output, None);
-    assert_eq!(state.selected_index(), Some(0));
-
-    // Expand should be ignored
-    let output = Tree::update(&mut state, TreeMessage::Expand);
-    assert_eq!(output, None);
-    assert_eq!(state.visible_count(), 1); // Should still be collapsed
-
-    // Select should be ignored
-    let output = Tree::update(&mut state, TreeMessage::Select);
-    assert_eq!(output, None);
-}
-
-#[test]
-fn test_update_expand_all_ignored_when_disabled() {
-    let mut root = TreeNode::new("Root", ());
-    root.add_child(TreeNode::new("Child", ()));
-
-    let mut state = TreeState::new(vec![root]);
-    state.set_disabled(true);
-
-    let output = Tree::update(&mut state, TreeMessage::ExpandAll);
-    assert_eq!(output, None);
-    assert_eq!(state.visible_count(), 1); // Should not have expanded
-}
-
-#[test]
-fn test_dispatch_event_ignored_when_disabled() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-    state.set_disabled(true);
-
-    let output = Tree::<&str>::dispatch_event(&mut state, &Event::key(KeyCode::Down));
-    assert_eq!(output, None);
-    assert_eq!(state.selected_index(), Some(0)); // Should not have moved
-}
-
-#[test]
-fn test_instance_handle_event_disabled() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-    state.set_disabled(true);
-
-    let msg = state.handle_event(&Event::key(KeyCode::Down));
-    assert!(msg.is_none());
-}
-
-#[test]
-fn test_instance_update_disabled() {
-    let mut state = make_tree_state();
-    state.set_disabled(true);
-
-    let output = state.update(TreeMessage::Down);
-    assert!(output.is_none());
-    assert_eq!(state.selected_index(), Some(0));
-}
-
-#[test]
-fn test_instance_dispatch_event_disabled() {
-    let mut state = make_tree_state();
-    state.set_focused(true);
-    state.set_disabled(true);
-
-    let output = state.dispatch_event(&Event::key(KeyCode::Down));
-    assert!(output.is_none());
-    assert_eq!(state.selected_index(), Some(0));
-}
-
-#[test]
-fn test_view_disabled() {
-    let root = TreeNode::new("Root", ());
-    let mut state = TreeState::new(vec![root]);
-    state.set_disabled(true);
-
-    let (mut terminal, theme) = crate::component::test_utils::setup_render(40, 10);
-
-    terminal
-        .draw(|frame| {
-            Tree::view(&state, frame, frame.area(), &theme);
-        })
-        .unwrap();
-
-    insta::assert_snapshot!(terminal.backend().to_string());
 }
