+++ src/component/tree/tests/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+use crate::input::{Event, KeyCode};
+
+mod component;
+mod events;
+mod node;
+mod state;
